move routes to csc context
